refactor: KDA-52 일기 수정 화면 패키지 이름 변경
