+++ presentation/src/main/java/com/upf464/koonsdiary/presentation/ui/main/diary/DiaryScreen.kt
@@ -6,9 +6,9 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.upf464.koonsdiary.presentation.common.Constants
-import com.upf464.koonsdiary.presentation.ui.main.diary.add.DiaryEditorScreen
 import com.upf464.koonsdiary.presentation.ui.main.diary.calendar.CalendarScreen
 import com.upf464.koonsdiary.presentation.ui.main.diary.detail.DiaryDetailScreen
+import com.upf464.koonsdiary.presentation.ui.main.diary.editor.DiaryEditorScreen
 
 @Composable
 internal fun DiaryScreen() {
