+++ presentation/src/main/java/com/upf464/koonsdiary/presentation/ui/main/diary/editor/DiaryEditorEvent.kt
@@ -1,4 +1,4 @@
-package com.upf464.koonsdiary.presentation.ui.main.diary.add
+package com.upf464.koonsdiary.presentation.ui.main.diary.editor
 
 sealed class DiaryEditorEvent {
